refactor: PopularThemeFinder 클래스 삭제
