--- src/main/java/roomescape/domain/ClockConfig.java
@@ -1,13 +1,0 @@
-package roomescape.domain;
-
-import java.time.Clock;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class ClockConfig {
-    @Bean
-    public Clock clock() {
-        return Clock.systemDefaultZone();
-    }
-}
